requirements.txt
# Core Data & Computation
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Financial Data Sources
yfinance==0.2.28
nsepython==2.5
pandas-datareader==0.10.0

# Options Pricing
mibian==0.1.3
py_vollib==1.0.1

# Dashboard
streamlit==1.28.0
plotly==5.17.0
streamlit-aggrid==0.3.4

# Database
sqlalchemy==2.0.21

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Logging
loguru==0.7.2